+++ tests/test_library.py
@@ -48,6 +48,15 @@ def test_get_images_for_unknown_uri_returns_empty_list(config, caplog):
     assert "Failed to lookup image for Pandora URI '{}'.".format(track_uri.uri) in caplog.text()
 
 
+def test_get_images_for_unsupported_uri_type_issues_warning(config, caplog):
+    backend = conftest.get_backend(config)
+
+    search_uri = PandoraUri.factory('pandora:search:R12345')
+    results = backend.library.get_images([search_uri.uri])
+    assert len(results[search_uri.uri]) == 0
+    assert "No images available for Pandora URIs of type 'search'.".format(search_uri.uri) in caplog.text()
+
+
 def test_get_images_for_track_without_images(config, playlist_item_mock):
     backend = conftest.get_backend(config)
 
@@ -75,7 +84,6 @@ def test_get_next_pandora_track_fetches_track(config, playlist_item_mock):
 
     station_mock = mock.Mock(spec=Station)
     station_mock.id = 'id_token_mock'
-    station_mock.id = 'id_token_mock'
     backend.library.pandora_station_cache[station_mock.id] = StationCacheItem(station_mock, iter([playlist_item_mock]))
 
     ref = backend.library.get_next_pandora_track('id_token_mock')
@@ -88,7 +96,6 @@ def test_get_next_pandora_track_handles_no_more_tracks_available(config, caplog)
     backend = conftest.get_backend(config)
 
     station_mock = mock.Mock(spec=Station)
-    station_mock.id = 'id_token_mock'
     station_mock.id = 'id_token_mock'
     backend.library.pandora_station_cache[station_mock.id] = StationCacheItem(station_mock, iter([]))
 
@@ -157,6 +164,31 @@ def test_lookup_of_ad_uri_defaults_missing_values(config, ad_item_mock):
     assert track.album.name == '(Company name not specified)'
 
 
+def test_lookup_of_search_uri(config, playlist_item_mock):
+    with mock.patch.object(MopidyAPIClient, 'get_station', conftest.get_station_mock):
+        with mock.patch.object(APIClient, 'create_station',
+                               mock.Mock(return_value=conftest.station_result_mock()['result'])) as create_station_mock:
+            with mock.patch.object(APIClient, 'get_station_list', conftest.get_station_list_mock):
+
+                backend = conftest.get_backend(config)
+
+                station_mock = mock.Mock(spec=Station)
+                station_mock.id = conftest.MOCK_STATION_ID
+                backend.library.pandora_station_cache[station_mock.id] = \
+                    StationCacheItem(conftest.station_result_mock()['result'],
+                                     iter([playlist_item_mock]))
+
+                track_uri = PlaylistItemUri._from_track(playlist_item_mock)
+                backend.library.pandora_track_cache[track_uri.uri] = TrackCacheItem(mock.Mock(spec=models.Ref.track),
+                                                                                    playlist_item_mock)
+
+                results = backend.library.lookup("pandora:search:S1234567")
+                # Make sure a station is created for the search URI first
+                assert create_station_mock.called
+                # Check that the first track to be played is returned correctly.
+                assert results[0].uri == track_uri.uri
+
+
 def test_lookup_of_track_uri(config, playlist_item_mock):
     backend = conftest.get_backend(config)
 
@@ -195,6 +227,19 @@ def test_lookup_of_missing_track(config, playlist_item_mock, caplog):
 
     assert len(results) == 0
     assert "Failed to lookup Pandora URI '{}'.".format(track_uri.uri) in caplog.text()
+
+
+def test_lookup_overrides_album_and_artist_uris(config, playlist_item_mock):
+    backend = conftest.get_backend(config)
+
+    track_uri = PlaylistItemUri._from_track(playlist_item_mock)
+    backend.library.pandora_track_cache[track_uri.uri] = TrackCacheItem(mock.Mock(spec=models.Ref.track),
+                                                                        playlist_item_mock)
+
+    results = backend.library.lookup(track_uri.uri)
+    track = results[0]
+    assert next(iter(track.artists)).uri == track_uri.uri
+    assert track.album.uri == track_uri.uri
 
 
 def test_browse_directory_uri(config):
@@ -329,6 +374,22 @@ def test_browse_station_uri(config, station_mock):
             assert len(results) == 1
 
 
+def test_formatted_search_query_concatenates_queries_into_free_text(config):
+    backend = conftest.get_backend(config)
+
+    result = backend.library._formatted_search_query({
+        'any': ['any_mock'], 'artist': ['artist_mock'], 'track_name': ['track_mock']
+    })
+    assert 'any_mock' in result and 'artist_mock' in result and 'track_mock' in result
+
+
+def test_formatted_search_query_ignores_unsupported_attributes(config):
+    backend = conftest.get_backend(config)
+
+    result = backend.library._formatted_search_query({'album': ['album_mock']})
+    assert len(result) is 0
+
+
 def test_refresh_without_uri_refreshes_root(config):
     backend = conftest.get_backend(config)
     backend.api.get_station_list = mock.Mock()
@@ -375,7 +436,6 @@ def test_refresh_station_directory(config):
 
     station_mock = mock.Mock(spec=Station)
     station_mock.id = 'id_token_mock'
-    station_mock.id = 'id_token_mock'
     backend.library.pandora_station_cache[station_mock.id] = StationCacheItem(station_mock, iter([]))
 
     backend.library.refresh('pandora:station:id_token_mock:id_token_mock')
@@ -393,3 +453,30 @@ def test_refresh_station_directory_not_in_cache_handles_key_error(config):
     assert backend.library.pandora_station_cache.currsize == 0
     assert not backend.api.get_station_list.called
     assert not backend.api.get_genre_stations.called
+
+
+def test_search_returns_empty_result_for_unsupported_queries(config, caplog):
+    backend = conftest.get_backend(config)
+    assert len(backend.library.search({'album': ['album_name_mock']})) is 0
+    assert 'Unsupported Pandora search query:' in caplog.text()
+
+
+def test_search(config):
+    with mock.patch.object(APIClient, 'search', conftest.search_mock):
+
+        backend = conftest.get_backend(config)
+        search_result = backend.library.search({'any': 'search_mock'})
+
+        assert len(search_result.tracks) is 2
+        assert search_result.tracks[0].uri == 'pandora:search:G123'
+        assert search_result.tracks[0].name == 'search_genre_mock (Pandora genre)'
+
+        assert search_result.tracks[1].uri == 'pandora:search:S1234567'
+        assert search_result.tracks[1].name == conftest.MOCK_TRACK_NAME + ' (Pandora station)'
+
+        assert len(search_result.artists) is 2
+        assert search_result.artists[0].uri == 'pandora:search:R123456'
+        assert search_result.artists[0].name == 'search_artist_artist_mock (Pandora artist)'
+
+        assert search_result.artists[1].uri == 'pandora:search:C123456'
+        assert search_result.artists[1].name == 'search_artist_composer_mock (Pandora composer)'